+++ bench/creation.js
@@ -49,6 +49,26 @@ tests["q({ num: bigint, den: bigint })"] = (name) => {
   console.timeEnd(name);
 };
 
+tests["q({ num: number, den: number })"] = (name) => {
+  const xs = pairs.map(([n, d]) => ({ num: n, den: d }));
+
+  console.time(name);
+  for (const x of xs) {
+    q(x);
+  }
+  console.timeEnd(name);
+};
+
+tests["q({ num: string, den: string })"] = (name) => {
+  const xs = pairs.map(([n, d]) => ({ num: String(n), den: String(d) }));
+
+  console.time(name);
+  for (const x of xs) {
+    q(x);
+  }
+  console.timeEnd(name);
+};
+
 tests['q("num/den")'] = (name) => {
   const xs = pairs.map(([n, d]) => `${n}/${d}`);
 
@@ -60,7 +80,7 @@ tests['q("num/den")'] = (name) => {
 };
 
 tests['q("num/1000")'] = (name) => {
-  const xs = pairs.map(([n, d]) => `${n}/1000`);
+  const xs = pairs.map(([n]) => `${n}/1000`);
 
   console.time(name);
   for (const x of xs) {